+++ flutter_local_device_discovery_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+* Updated version to 0.2.0 for federated plugin alignment.
+
 ## 0.1.0
 
 * Initial release of the Windows implementation of flutter_local_device_discovery.